Kernel-weighted predictions from regression models

Calculates kernel-weighted predictions from regression models (i.e. outcomes that can be calculated from the predict function).

sm_predict(data, method, formula, type, newdata = data,
  method.args = NULL, lambda = 1, kernel = "epanechnikov",
  dist.method = "euclidean", verbose = FALSE)

Arguments

data
data frame

method
function to use

formula
formula

type
type of statistic to smooth (e.g. survival, median survival, etc.)

newdata
new data frame. Default is `data`. Only requires covariates from the RHS of `~` and the time component from the outcome for some survival estimators.

method.args
List of additional arguments passed on to the modelling function defined by `method`

lambda
The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels. The standard deviation for the Gaussian kernel

kernel
Specifies the kernel to be used: `epanechnikov`, `tricube`, `gaussian`, and `flat` are accepted. Default is `epanechnikov`

dist.method
Specifies the distance measure to be used in the kernel. Default is `euclidean`. Distance measures accepted by

verbose
Return full set of results as an attribute. Default is `FALSE`

Examples

sm_predict(
  mtcars,
  method = "glm",
  formula = am ~ mpg,
  method.args = list(family = binomial(link = "logit")),
  type = "response"
)

#> Warning in glm: non-integer #successes in a binomial glm!

#> # A tibble: 27 x 2
#>       am   mpg
#>    <dbl> <dbl>
#>  1     1  21  
#>  2     1  22.8
#>  3     0  21.4
#>  4     0  18.7
#>  5     0  18.1
#>  6     0  14.3
#>  7     0  24.4
#>  8     0  22.8
#>  9     0  19.2
#> 10     0  17.8
#> 11     0  16.4
#> 12     0  17.3
#> 13     0  15.2
#> 14     0  10.4
#> 15     0  14.7
#> 16     1  32.4
#> 17     1  30.4
#> 18     1  33.9
#> 19     0  21.5
#> 20     0  15.5
#> 21     0  13.3
#> 22     1  27.3
#> 23     1  26  
#> 24     1  15.8
#> 25     1  19.7
#> 26     1  15  
#> 27     1  21.4

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb      .fitted
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 4.089578e-01
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 4.089578e-01
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 5.504164e-01
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 4.370929e-01
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 3.016828e-01
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 2.728354e-01
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 1.461361e-01
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 6.456536e-01
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 5.504164e-01
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 3.238854e-01
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 2.580795e-01
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 1.993910e-01
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 2.351012e-01
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 1.621707e-01
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 1.057911e-05
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 1.057911e-05
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 1.535469e-01
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 1.000000e+00
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 1.000000e+00
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 1.000000e+00
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 4.449265e-01
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 1.685713e-01
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 1.621707e-01
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 1.201609e-01
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 3.238854e-01
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 9.454960e-01
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 7.926169e-01
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 1.000000e+00
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 1.774318e-01
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 3.448688e-01
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 1.590070e-01
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 4.370929e-01