Kernel-weighted regression modelling

General function that fits weighted regerssion models, where the weights are calculated from a ancillary variable(s) or from variable(s) found in the regression model. Using `data`, `method`, and `formula` the regression model is estimated; the estimates are weighted by the variable(s) listed in `weighting_var`.

sm_regression(data, method, formula, weighting_var, newdata = data,
  method.args = NULL, lambda = 1, kernel = "epanechnikov",
  dist.method = "euclidean")

Arguments

data
data frame

method
function to use

formula
formula

weighting_var
columns name(s) of variables used to calculate weights

newdata
new data frame. Default is `data`.

method.args
List of additional arguments passed on to the modelling function defined by `method`

lambda
The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels. The standard deviation for the Gaussian kernel

kernel
Specifies the kernel to be used: `epanechnikov`, `tricube`, `gaussian`, and `flat` are accepted. Default is `epanechnikov`

dist.method
Specifies the distance measure to be used in the kernel. Default is `euclidean`. Distance measures accepted by stats::dist is acceptable.

Examples

sm_regression(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  weighting_var = "hp",
  lambda = 2
)

#> # A tibble: 31 x 9
#>    newdata newdata_scaled distance weight model_safely model_error model_warning
#>    <list>  <list>         <list>   <list> <list>       <list>      <list>       
#>  1 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  2 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  3 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  4 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  5 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  6 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  7 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  8 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#>  9 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#> 10 <tibbl~ <tibble [1 x ~ <dbl [3~ <dbl ~ <list [2]>   <NULL>      <chr [0]>    
#> # ... with 21 more rows, and 2 more variables: model_message <list>,
#> #   model_obj <list>